LPS-102185 Move configuration and constants to API module
