+++ modules/apps/portal-security-sso/portal-security-sso-token-api/build.gradle
@@ -1,3 +1,5 @@
 dependencies {
+	compileOnly group: "biz.aQute.bnd", name: "biz.aQute.bnd.annotation", version: "4.2.0"
 	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
+	compileOnly project(":apps:static:portal-configuration:portal-configuration-metatype-api")
 }
