+++ modules/apps/portal-security-sso/portal-security-sso-token-api/src/main/resources/com/liferay/portal/security/sso/token/configuration/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
